feat: add autofinish
