+++ src/senders/finish.sender.ts
@@ -1,19 +1,8 @@
+import { getFinishResponse } from '../services/finish.service';
 import { log } from '../services/log.service';
-import { TypeFinishResponse } from '../types/finish.type';
 import { TypeShipData } from '../types/ship.type';
-import { CommandTypes, ID_VALUE } from '../utils/constants';
+import { CommandTypes } from '../utils/constants';
 import { searchWinner } from './updateWinners.sender';
-
-export const getFinishResponse = (idPlayer: number): string => {
-  const response: TypeFinishResponse = {
-    type: CommandTypes.finish,
-    data: JSON.stringify({
-      winPlayer: idPlayer,
-    }),
-    id: ID_VALUE,
-  };
-  return JSON.stringify(response);
-};
 
 export const finish = (data: TypeShipData[] | undefined, idPlayer: number) => {
   if (data) {
